# Speed up `list_resources` function for large collections