--- a/nmdc_runtime/api/endpoints/util.py
+++ b/nmdc_runtime/api/endpoints/util.py
@@ -72,20 +72,17 @@ def does_collection_contain_more_than_n_matching_documents(
     Inspired by: https://stackoverflow.com/a/67503437
     """
 
-    # Ensure the `n` value is non-negative, since the aggregation pipeline's
-    # `$limit` stage requires a non-negative integer.
-    if n < 0:
-        raise ValueError("The `n` value must be non-negative.")
+    # Ensure the `n` value is at least 0, since the aggregation pipeline's
+    # `$limit` stage requires a positive integer.
+    if not n >= 0:
+        raise ValueError("The `n` value must be at least 0.")
 
-    result: List[Dict[str, int]] = list(
-        collection.aggregate(
-            [
-                {"$match": filter_},
-                {"$limit": n + 1},
-                {"$count": "numCounted"},
-            ]
-        )
-    )
+    result_cursor = collection.aggregate([
+        {"$match": filter_},
+        {"$limit": n + 1},
+        {"$count": "numCounted"},
+    ])
+    result: List[Dict[str, int]] = list(result_cursor)
 
     # Note: If no documents match the filter, `result` will be an empty list;
     #       otherwise, it will be a 1-item list consisting of a dictionary.