<!--┌─────────────────────────────────────────────────────────────────────────┐
    │ 1. Summary (required)                                                   │
    │                                                                         │
    │ Summarize the changes you made on this branch. This is typically a more │
    │ detailed restatement of the PR title.                                   │
    │                                                                         │
    │ Example: "On this branch, I updated the `/studies/{study_id}` endpoint  │
    │           so it returns an HTTP 404 response when the specified study   │
    │           does not exist."                                              │
    └─────────────────────────────────────────────────────────────────────────┘-->

On this branch, I updated the `list_resources` function so that, instead of counting _all_ matching documents in a collection, it only counts _the minimum number necessary_ in order to do its job.

Here's an analogy:

<img width="462" height="243" alt="image" src="https://github.com/user-attachments/assets/dbce7f2b-d06f-48af-baca-6627b5c7b109" />

I have a yard with lots of different types of trees in it. I want to know whether my yard has **more than 3** oak trees. I can either:
1. (the old way) Ask someone to **count all** the oak trees in my yard, and tell me that number (I'll compare it with 3); or
2. (the new way) Ask someone to **count up to 4** oak trees in my yard, and tell me that number (I'll compare it with 3)

### Details

<!--┌─────────────────────────────────────────────────────────────────────────┐
    │ 2. Details (optional)                                                   │
    │                                                                         │
    │ Provide additional information you think readers will find useful.      │
    │ Readers include PR reviewers, release note authors, app debuggers, and  │
    │ your future self. Additional information might include motivation,      │
    │ rationale, and a description of how things used to be.                  │
    │                                                                         │
    │ Example: "It previously returned an HTTP 404 response and an empty      │
    │           JSON object."                                                 │
    └─────────────────────────────────────────────────────────────────────────┘-->

As a result, the `GET /nmdc_schema/{collection_name}` endpoint (and other endpoints and functions that use the `list_resources` function under the hood) is faster. See load testing results and performance profiling results below for information about "how much faster."

#### Load testing results

| Before<br/>(i.e. commit 722c3f6073312dc141abc9349efe03bbefadfd24, tip of `main` branch) | After<br/>(i.e. commit https://github.com/microbiomedata/nmdc-runtime/pull/1166/commits/7789b00e96742d4befe3a0b978efc0bf99da0c23, near tip of PR branch) |
| --- | --- |
| <img width="400" alt="image" src="https://github.com/user-attachments/assets/a5deebf9-2c16-42af-b8ac-ee809f34fb1b" /> | <img width="400" alt="image" src="https://github.com/user-attachments/assets/2381bd09-727e-4b27-bbda-6a2a90ea0431" /> |
| Median response time for<br/>`GET /nmdcschema/functional_annotation_agg`<br/>under the test load: 76 seconds | Median response time for<br/>`GET /nmdcschema/functional_annotation_agg`<br/>under the test load: 300 milliseconds (~250x faster) | 


<details>
<summary>Show/hide load testing notes</summary>

Here's the command I used to launch the load test(er):

```sh
docker run --rm -it \
    --publish 8089:8089 \
    --volume ${PWD}/util/load_testing:/mnt/locust \
    --network nmdc-runtime-dev_default \
    locustio/locust \
      --locustfile /mnt/locust/locustfile.py \
      --host http://fastapi:8000 --autostart --users 25 --spawn-rate 5 -t 30s User --stop-timeout 180s
```

Notes:
- In the "Before" case, I used the `locustfile.py` file as it existed in the "After" case (i.e. same load configuration in both cases).
- I included `--stop-timeout 180s` in the command because—without it—in the "Before" case, the HTTP responses to the `GET /nmdcschema/functional_annotation_agg` requests were taking so long to come back (about 2 minutes, although that's dependent upon my local environment) that the load tester client would not acknowledge them at that point anymore. When using this 3-minute cleanup period, the load tester client does acknowledge them.

</details>

#### Performance profiling results

Here's how this change impacts the `GET /nmdcschema/functional_annotation_agg` endpoint's performance:

| Before<br/>(i.e. commit 722c3f6073312dc141abc9349efe03bbefadfd24, tip of `main` branch) | After<br/>(i.e. commit https://github.com/microbiomedata/nmdc-runtime/pull/1166/commits/7789b00e96742d4befe3a0b978efc0bf99da0c23, near tip of PR branch) |
| --- | --- |
| <img width="400" alt="image" src="https://github.com/user-attachments/assets/7c773027-d9af-4349-8a91-99f8434fb234" /> | <img width="400" alt="image" src="https://github.com/user-attachments/assets/e3f517d9-2d4d-4721-999f-cf0c5056c237" /> |
| Response time: ~20 seconds | Response time: ~20 milliseconds (~1000x faster) |

<details>
<summary>Show/hide performance profiling notes</summary>

Notes:
- In both the "Before" and "After" cases, I prepended `async` to the definition of the endpoint function.
  ```diff
    @router.get(
        "/nmdcschema/{collection_name}",
        response_model=ListResponse[Doc],
        response_model_exclude_unset=True,
    )
  - def list_from_collection(
  + async def list_from_collection(
  ```
- The performance profiling was performed when the Runtime was not receiving any requests _other than_ the one being profiled.

</details>

Also on this branch:
- I added a check for `max_page_size` being something other than an integer, since its type hints say it can be `int | None`.
- I refactored some code in an attempt to make the conditions under which pagination will be performed more apparent to maintainers of this codebase.
- I refined the load testing instructions and related configuration file, based upon my experience measuring the performance of the `GET /nmdcschema/functional_annotation_agg` endpoint. I also configured Uvicorn's file watcher to ignore those files, since they aren't part of the FastAPI application.

### Related issue(s)

<!--┌─────────────────────────────────────────────────────────────────────────┐
    │ 3. Related issue(s) (optional)                                          │
    │                                                                         │
    │ Link to any GitHub issue(s) this branch was designed to resolve.        │
    │                                                                         │
    │ Example: "Fixes #12345"                                                 │
    └─────────────────────────────────────────────────────────────────────────┘-->

Fixes #1164 

### Related subsystem(s)

<!--┌─────────────────────────────────────────────────────────────────────────┐
    │ 4. Related subsystem(s) (required)                                      │
    │                                                                         │
    │ Mark the checkbox next to each subsystem related to the changes in this │
    │ branch. This information might influence who you request reviews from.  │
    │                                                                         │
    │ Example: If you modified the `/studies/{study_id}` API endpoint,        │
    │          mark the checkbox next to "Runtime API (except the Minter)".   │
    └─────────────────────────────────────────────────────────────────────────┘-->

- [x] Runtime API (except the Minter)
- [ ] Minter
- [ ] Dagster
- [ ] Project documentation (in the `docs` directory)
- [ ] Translators (metadata ingest pipelines)
- [ ] MongoDB migrations
- [ ] Other

### Testing

<!--┌─────────────────────────────────────────────────────────────────────────┐
    │ 5. Testing (required)                                                   │
    │                                                                         │
    │ Indicate whether you have already tested the changes this branch        │
    │ contains; and, if so, how someone other than you can test them. That    │
    │ may involve attaching example files or ad hoc test instructions.        │
    │                                                                         │
    │ Example: "I tested these changes by adding a pytest test that ensures   │
    │           the database does not contain a Study whose ID is `foo`,      │
    │           then submits an HTTP request to `/studies/foo` and confirms   │
    │           the response status is 404."                                  │
    └─────────────────────────────────────────────────────────────────────────┘-->

- [x] I tested these changes (explain below)
- [ ] I did not test these changes

The test I added, which targets the newly-added helper function, passes locally and on GHA. I also performed load testing and performance profiling (shown above).

### Documentation

<!--┌─────────────────────────────────────────────────────────────────────────┐
    │ 6. Documentation (required)                                             │
    │                                                                         │
    │ Indicate whether, in this branch, you have updated all documentation    │
    │ that would otherwise become inaccurate if this branch were to be        │
    │ merged in.                                                              │
    └─────────────────────────────────────────────────────────────────────────┘-->

- [x] I **have not checked** for relevant documentation yet (e.g. in the `docs` directory)
- [ ] I have **updated** all relevant documentation so it will remain accurate
- [ ] Other (explain below)

### Maintainability

<!--┌─────────────────────────────────────────────────────────────────────────┐
    │ 7. Maintainability (required)                                           │
    │                                                                         │
    │ Indicate whether you have done each of these things that can make code  │
    │ easier to maintain, whether by your teammates or by your future self.   │
    └─────────────────────────────────────────────────────────────────────────┘-->

- [x] Every Python function I defined includes a docstring _(test functions are exempt from this)_
- [x] Every Python function parameter I introduced includes a type hint (e.g. `study_id: str`)
- [x] All "to do" or "fix me" Python comments I added begin with either `# TODO` or `# FIXME`
- [ ] I used `black` to format all the Python files I created/modified
- [x] The PR title is in the imperative mood (e.g. "Do X") and not the declarative mood (e.g. "Does X" or "Did X")