@eecavanna 2025-08-24
I just noticed Mongo's `collection.countDocuments` function accepts a `limit` option.

https://www.mongodb.com/docs/manual/reference/method/db.collection.countDocuments/#syntax

If pymongo's `collection.count_documents` function supports that, then we could use that instead and avoid having this helper function I am introducing in this PR (that could be addressed now or via a future, refactoring PR).

https://pymongo.readthedocs.io/en/stable/api/pymongo/collection.html#pymongo.collection.Collection.count_documents